Updated the cheetah and pdsdata libraries so that it is compatible with ConfigV4 of the CSPAD. Current version of release: 2012-11-28

+++ release/pdsdata/ipimb/src/ConfigV2.cc
@@ -18,8 +18,8 @@ ConfigV2::~ConfigV2() {}
 
 void ConfigV2::dump() const {
   printf("------Ipimb Config-------------\n");
-  printf("Trigger counter 0x%llx\n", triggerCounter());
-  printf("Serial ID 0x%llx\n", serialID());
+  printf("Trigger counter 0x%llx\n", (long long) triggerCounter());
+  printf("Serial ID 0x%llx\n", (long long) serialID());
   printf("Charge amp settings 0x%x\n", chargeAmpRange());
   printf("Reset length %ld ns\n", (unsigned long)resetLength());
   printf("Reset delay %d ns\n", resetDelay());
